~/projects/tennis-betting-system/requirements.txt
# Core
python-dotenv==1.0.0
requests==2.31.0

# Scraping
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1

# Notion
notion-client==2.2.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0

# Monitoring
sentry-sdk==1.38.0

